Add files via upload
This lab was created for me to learn more about Active Directory and how it would be used in a working environment.  I was able to learn a lot about how important Active Directory is in an enterprise. I learned to create and populate a server and create policies within that server for users.